[CodeBuild] changes from code review, security update
